frontend/requirements.txt
# Core Framework
streamlit==1.40.2

# HTTP Client
requests==2.32.3

# Utilities
python-dotenv==1.0.1

# UI Enhancement
streamlit-javascript==0.1.5
streamlit-extras==0.4.7
